address statistic

+++ education/education-common/src/main/java/com/ssic/education/common/dto/AddressStatistic.java
@@ -55,7 +55,7 @@ public class AddressStatistic {
 	 */
 	@Getter
 	@Setter
-	private Float schoolNum;
+	private Integer schoolNum;
 	
 }
